Fixed VS2010 warnings

+++ CORE/coordinates.h
@@ -75,7 +75,7 @@ public:
 	CoordinatesF() {_x = 0; _y = 0;}
 	CoordinatesF(float x, float y) {_x = x; _y = y;}
 	CoordinatesF(const CoordinatesF &c) {_x = c.x(), _y = c.y();}
-	CoordinatesF(const Coordinates &c) {_x = c.x(); _y = c.y();}
+	CoordinatesF(const Coordinates &c) {_x = (float)c.x(); _y = (float)c.y();}
 
 	float x() const {return _x;}
 	float y() const {return _y;}
